extracted replay logic into a custom hook
